fix: work with swift 5.4

+++ Wonky Blocks/UI/GameViewController.swift
@@ -46,7 +46,7 @@ class WonkyGameViewController: UIViewController, SKSceneDelegate {
     lastUpdateTime = currentTime
 
     if !gameState.paused {
-      physicsController.handleKeyEvents(delta: delta)
+      physicsController.handleKeyEvents(delta: CGFloat(delta))
     }
   }
